将参数请求全部改为类请求
包括三大列表控制器

+++ src/main/java/com/wuyanteam/campustaskplatform/entity/UTT.java
@@ -26,7 +26,7 @@ public class UTT
     private int likeCount;
     private String realName;
     private String address;
-    private int balance;
+    private float balance;
     private Timestamp lastLoginTime;
     private int takeNum;
     private int publishNum;
@@ -42,7 +42,7 @@ public class UTT
     private Timestamp publishTime;
     private String state;
     private Timestamp takeTime;
-    private int reward;
+    private float reward;
     private String startAddress;
     private String endAddress;
     private Timestamp dueTime;
